refactor(fig): use `stepping` variable instead of mutating options
This is almost the only place we were mutating options. By moving the
mutation into a local variable we can treat `options` as static and get
rid of the notion of `currentOptions`, which in turn will make it easy
to run aspects in parallel.

The other place we're doing some mutation is in this nested object:

  options.startAt.found // boolean
  options.startAt.literal // string

This happens during the "pre-flight" phase before there is any
parallelism, so this is pretty innocuous.