refactor(fig): teach `VariableRegistry` to handle static variables too
Well, not _totally_ static, but mostly so. Prior to this commit,
`VariableRegistry` only existed to support aspects using the
`variables()` DSL like this to dynamically derive more variables at
runtime (example from "dotfiles" aspect):

    variables(({hostHandle, identity, profile}) => {
      return {
        gitGpgSign: identity === 'wincent' && profile !== 'codespaces',
        gitHostSpecificInclude: `.gitconfig.d/${hostHandle}`,
        gitUserEmail: identity === 'wincent' ? 'greg@hurrell.net' : '',
        gitUserName: identity === 'wincent' ? 'Greg Hurrell' : '',
        gitHubUsername: identity === 'wincent' ? 'wincent' : '',
      };
    });

Maybe this whole API is overkill, who knows? That's the only place I'm
using it. But the basic idea is to take in the currently computed
variables and give the aspect a chance to derive something additional at
runtime.

With this commit, we get the `VariableRegistry` to _also_ remember all
the other variables. We call these "static" just because that's a
convenient complement to "dynamic", but strictly speaking they are _not_
totally static:

- Base: ie. `profile`
- Attributes: ie. static system-related properties.
- Defaults: defined in project.json
- Profile: profile-level overrides defined in project.json
- Platform: platform-level overrides defined in project.json
- Dynamic: computed in variables.ts
- Aspect (static): defined in "aspect/*/aspect.{json,ts}".
- Aspect (derived): defined using `variables()` DSL.

So... now that I write that out I realize that "dynamic" is already
being used fo the 6th group above, and in a way, we have multiple layers
of "derived" variables so that's not really a great name either.
Whatever, I'll think a bit more about it and see if I can come up with
better names.

Regardless, the motivation here is noted in the inline comments: I want
to get away from undifferentiated global state and organize variables by
aspect so that I can run aspects in parallel. This is an incremental
step in that direction.