refactor(fig): rename `getCaller()` to `getCallers() etc
Instead of looking for a direct caller in the third slot of the stack
trace, look through the whole trace. This means that instead of just
some select places like inside `task()` and `handler()`, we can in
theory look up our aspect from just about anywhere; eg. in a `command()`
call, we should be able to find the aspect out, which is a necessary
precondition for being able to dispense with our naughty shared
singleton state.