fix(fig): fix event handling in `run()` wrapper
So, looks like I was relying on events coming back in a certain order in
such a way that was fine when running serially, but which exposed me to
bugs when running multiple child processes concurrently, as seen in the
parent commit.

Careful consultation of the NodeJS docs reveals that I should be able to
rely on getting a `close` event after the child process has exited (ie.
after the `exit` event) or after it otherwise fails (ie. the `error`
event).

So, this commit moves the `resolve()` call into the `close` handler,
where it seems that it always fires reliably.