feat(fig): add -v/--verbose flag
Although it is admittedly kind of pointless because we run at a log
level of 6 ("NOTICE") by default, and `--debug` bumps us up to 7
("DEBUG") anyway, which is the max.

This is fruit of using syslog levels as a basis: most of the levels
aren't really interesting in a system like Fig, in the sense that level
3 ("ERROR") is already bad enough to abort the entire process, so the
lower levels (2 = "CRITICAL", 1 = "ALERT", 0 = "EMERGENCY") are all
basically useless.

In a future refactoring I might move away from syslog levels so that we
have something more useful in between "INFO" and "DEBUG".