perf(darwin): run some more things in parallel
Basically, the things I am _not_ running in parallel yet are things
which might produce a `sudo` prompt, seeing as I don't have a mutex in
place yet to stop prompts from getting jumbled in with other output.

On an already set-up machine, most tasks are skipped or already ok, so
there's no real performance difference between `--parallel`:

    [notice]  Summary: changed=39 failed=0 ok=196 skipped=47 elapsed=30.51s

and without it:

    [notice]  Summary: changed=39 failed=0 ok=196 skipped=47 elapsed=30.13s

I expect the difference to actually be notable when real I/O work starts
happening. I'll test that out on a Codespace soonish.

Addendum: the above times are with the "homebrew" aspect temporarily
removed, because that one is always dog slow.