refactor: implement console mutex
So that even when `--parallel` is in effect, we won't wind up emitting
log lines at the same time as trying to prompt.

This isn't perfect because two aspects could still race to prompt the
user (eg. for sudo credentials) and the user would get prompted twice in
a row, but the odds of such a collision are pretty low and the downsides
are not exactly deal-breakers (everything will still work).

The cost is a lot more functions need to be async now, because we have
to await in lots more places. And we have to be careful to emit related
log lines together in one go instead of in multiple calls; eg. in
`log()` itself, we can no longer call `print()` twice to print a message
then a newline, but instead have to make one call.