chore: un-parallelize "darwin" tasks
Now that the parallel stuff is all tested and "proven", I am keeping it
where it matters (in the "codespaces" platform, to speed up spinning up
a new instance) and removing it where it isn't really necessary (eg.
"darwin", where I care more about easy-to-read, serial output).

Note that because `--parallel` is not in effect by default, this is
really a no-op change for most runs because things were going to run
serially anyway. But I think it makes it clearer if I just spell all
this out serially in the project manifest.