chore: update type definitions
New definitions reflect that `getgid()` is only available on POSIX
platforms, leading to these type errors:

    fig/Attributes.ts:101:19 - error TS2532: Object is possibly 'undefined'.

    101       this.#gid = process.getgid();
                          ~~~~~~~~~~~~~~

    fig/Attributes.ts:101:19 - error TS2722: Cannot invoke an object which is possibly 'undefined'.

    101       this.#gid = process.getgid();
                          ~~~~~~~~~~~~~~

    Found 2 errors in the same file, starting at: fig/Attributes.ts:101

As you can see in the diff, `getuid()` already had a type definition
like that, hence my `typeof` check. I refactored that to make it match
the change I'm making to the `getgid()` call as well.

The explicit imports of `process` aren't actually needed, but it seemed
a nice idea to make the dependence on the global explicit, so I went
through and did that too.