chore(ssh): add notes on `TCPKeepAlive` etc to Codespaces config
From `man ssh_config`:

     TCPKeepAlive
             Specifies whether the system should send TCP keepalive
             messages to the other side.  If they are sent, death of the
             connection or crash of one of the machines will be properly
             noticed.  However, this means that connections will die
             if the route is down temporarily, and some people find it
             annoying.

             The default is yes (to send TCP keepalive messages), and
             the client will notice if the network goes down or the
             remote host dies.  This is important in scripts, and many
             users want it too.

             To disable TCP keepalive messages, the value should be set
             to no.  See also ServerAliveInterval for protocol-level
             keepalives.

I have done a fair bit of connection troubleshooting over the last week
or so, and was wondering if this might help. As the docs say, when this
is `yes` (the default), "connections will die if the route is down
temporarily". I am not sure how long a route can be down "temporarily"
without it actually being fatal to the connection, but it wouldn't seem
to hurt to turn this to `no` in case it does help in some rare cases.

In reality, however, there is _no_ value in changing this setting,
because of the way codespaces work: you run `gh cs ssh` and what
actually happens under the covers is your `ssh` command connects to
_localhost_ (that connection is never going to die under normal
circumstances) and `gh` takes care of proxying the actual traffic via
magical means. So in this commit, I'm just putting a comment in the
config explaining all of this.

In any case, I don't need anything special to notice when one of these
connections is down anyway: I'm usually actively using the SSH session
interactively if I care. For this reason, there's _also_ no benefit in
turning on `ServerAliveInterval` (defaults to 0) either. Leaving these
at their defaults, which means they do nothing:

     ServerAliveCountMax
             Sets the number of server alive messages (see below) which
             may be sent without ssh(1) receiving any messages back from
             the server.  If this threshold is reached while server
             alive messages are being sent, ssh will disconnect from the
             server, terminating the session.  It is important to note
             that the use of server alive messages is very different
             from TCPKeepAlive (below).  The server alive messages are
             sent through the encrypted channel and therefore will
             not be spoofable.  The TCP keepalive option enabled by
             TCPKeepAlive is spoofable.  The server alive mechanism is
             valuable when the client or server depend on knowing when a
             connection has become unresponsive.

             The default value is 3.  If, for example,
             ServerAliveInterval (see below) is set to 15 and
             ServerAliveCountMax is left at the default, if the
             server becomes unresponsive, ssh will disconnect after
             approximately 45 seconds.

     ServerAliveInterval
             Sets a timeout interval in seconds after which if no data
             has been received from the server, ssh(1) will send a
             message through the encrypted channel to request a response
             from the server.  The default is 0, indicating that these
             messages will not be sent to the server.