feat(dotfiles): teach `ts`, `tw`, `tick` to show human-readable dates too
So we take a Git-internal date like:

    1658346637 +0200

(ie. UTC UNIX Epoch timestamp plus numeric time zone offset) and with it
split into two variables:

    TS = 1658346637
    TZ = +0200

We make a human readable form with:

    $TS TZ * 36
    1658346637 +0200 * 36
    1658346637 + 7200
    1658353837

And feed that to `strftime %c`, which will print a "preferred
representation for the current locale". For me, on this machine, that
ends up being:

    Wed 20 Jul 2022 21:50:37