chore(nvim): point vim-clipper back at `main` (from `next`)
As noted here:

    https://github.com/wincent/clipper/issues/26

I still don't have a good story for making Clipper do its thing both
locally (when I am sitting in front of my Linux box) and remotely (when
I am SSH'd into it from my laptop). As that issue describes, you _can_
use the OSC 52 support in kitty to get stuff into the clipboard
transparently, but, as explained here:

    https://github.com/kovidgoyal/kitty/issues/3937

it has a hard cap on the size that you can copy this way, which makes me
hesitant to fully rely on it.

You can customize the limit with `clipboard_max_size`, but the fact that
there is a limit at all leads me to analysis paralysis... how to set
an arbitrary quantity like that? And even if I do get everything working
with kitty, I'm only using that on my personal machines, not my work
machine, so if I ever need to SSH in from my work machine to the Linux
box to do something (eg. grab some changes to my dotfiles or something)
I'll be left with inconsistent support.

I've gone back and forth on this in the past, see:

    commit bf75f40bdc2f2f9a0343ac422e4a61c1542e96c6
    Date:   Sun Aug 15 10:57:49 2021 +0200

        refactor(dotfiles): [demo] teach `clip` to use OSC 52 escape sequence

    commit ef94ece8ee3ff2591bde4fc7b1179d080d043d14
    Date:   Sun Aug 15 11:02:57 2021 +0200

        chore: revert "refactor(dotfiles): [demo] teach `clip` to use OSC 52 escape sequence"

    commit c85890b0dada7ad1611501b7cde3a904aa48764c
    Date:   Sun Aug 15 11:17:03 2021 +0200

        refactor(dotfiles): switch back to using OSC 52 for `clip` when in Kitty