chore(nvim): set 'ignorecase'/'smartcase'
It's true that Loupe will set these for us later, but I want them to be
set at the time I make the call to `commandt.setup()`, so that it can
set itself accordingly. I could also pass in an explicit:

    ignore_case = true,
    smart_case = true,

But as the pythonistas say, implicit is better than implicit.

Or do they?