feat(fig): add the ability to skip aspects with "^" or "!" syntax
As this only took a few minutes to fix, decided to go ahead and do it
now.

Common use cases are:

    ./install dotfiles  # install one aspect
    ./install           # install all aspects
    ./install ^homebrew # install all but one aspect

but, technically, you can use "^" and "!" arbitrarily to mean "but not
this one". That means you can do nonsense like:

    ./install dotfiles ^dotfiles

which installs nothing at all. Note that a subsequent inclusion does not
trump a prior exclusion; exclusion is definitive; ie:

    ./install ^dotfiles dotfiles dotfiles dotfiles

also install nothing.

Closes: https://github.com/wincent/wincent/issues/127