fix(nvim): make focus restoration work for buffers without a filetype
Our logic here was bad. The intent was obviously "only do a table
look-up if we have a non-nil key", but the effect was "fail to restore
focus if we don't have a filetype".

A slight tweak to the logic gives us the right semantics: "restore focus
by default, unless the filetype tells you not to".