feat: guard against numeric hostname
No idea why this happened, but on my work Mac, `hostname` in the shell
and `require('os').hostname()` in NodeJS were both returning a numeric
IP address. This in turn produced a host handle of "192" which was
obviously useless, and broke a bunch of stuff (like, for example, I
couldn't make Git commits!).

    https://apple.stackexchange.com/questions/30552/os-x-computer-name-not-matching-what-shows-on-terminal

Goes to great lengths offering possible solutions here, and advice on
what you should and shouldn't do. Whatever, in my case the "Sharing"
panel showed the name, but:

```
scutil --get ComputerName     # showed the name...
scutil --get LocalHostName    # showed the name...
scutil --get HostName         # did NOT show the name...
```

`man hostname` says to make hostname changes permanent across reboots
with `sudo scutil --set Hostname $NAME`, despite the advice in the Stack
Exchange thread, so I went ahead and did that and everything is working
now.

In any case, adding some logging so that I'll hopefully notice this if
it happens again.