feat(dotfiles): teach `git ff` to autostash
As noted here:

    https://stackoverflow.com/a/30209750/2103996

there are myriad options and some can be annoying dangerous (just as
_not_ using the options can be annoying).

Seems like adding an explicit `--autostash` to `git ff` is fine though.
I can guarantee you that if I am trying to run `git ff` and it doesn't
work because of dirty state, I am guaranteed to run `git stash` as my
very next command. Often, the stash pops off cleanly after that (but
even if it didn't, I'd still want to try). So this seems like it
will likely be a time-saver for me.