chore(pacman): add evtest
Was just working on a change to the "interception" aspect to pass
certain combos through as function keys so that they can be used
(abused) as dead-keys in `~/.XCompose` to give me a few more Mac-like
binding on Linux.

`evtest` helpfully verifies that the right keycodes are coming through,
although the XCompose layer itself sees them as something else (ie.
`Pause` works fine, but `F22` does not). For these higher-order keys, it
seems that `evtest` shows things accurately in a way that `xev` does
not. eg.

- `evtest` shows `F23`, but `xev` shows `XF86TouchpadOff` (interception
  tools is sending F23).
- Ditto for `F22` (`evtest`) and `XF86TouchpadOn` (`xev`).