refactor(dotfiles): add .Xmodmap
Figured out that the reason why the kernel is sending F21/F22/F23 but
some part of the X stack is instead seeing
XF86TouchpadToggle/XF86TouchpadOn/XF86TouchpadOff is because of mappings
defined in /usr/share/X11/xkb/symbols/inet (excerpt):

    // Evdev Standardized Keycodes
    partial alphanumeric_keys
    xkb_symbols "evdev" {
        key <FK21>   {      [ XF86TouchpadToggle    ]       };
        key <FK22>   {      [ XF86TouchpadOn        ]       };
        key <FK23>   {      [ XF86TouchpadOff       ]       };

So, I could leave everything as-is, but I think I prefer to do the
override here. That way, `evtest` and `xev` see the same keys, and the
whole thing is lot more coherent (all the way from the kernel key codes,
via Interception Tools, up through the X stack, they all agree on what
the keys are), even though it does mean adding another file and having
to use it in two places. Part of the ugliness here is that we have two
systems interacting: the "legacy" `xmodmap` layer and the "modern"
`setxkbmap` layer, so it all feels a bit fragile and prone to change;
but with any luck, it will probably keep working for years.