fix(dotfiles): make alt+click open links in Kitty consistently
I feel like this used to work consistently on both macOS and Linux, but
maybe I am misremembering: one thing I am clear on is that before this
commit, my old habit of alt-clicking URLs in Kitty was no longer opening
them in the browser on either Linux or macOS.

This this change, it seems to work in both places again.

Note, as the comment in the commit describes, I'm using "alt" here in
the Linux sense. On Linux "alt-click" means holding down the big
modifier next to the space bar and clicking. On macOS, we tell Kitty to
"cmd-click" instead, which makes it do the same thing (open the link in
the browser when you click on it while holding the big modifier).