feat(tampermonkey): add script to unbreak YouTube
I'm in some A/B test (at least, I only see this when logged in), but
there is some new (as in, within the last month or two) design on
YouTube that prevents me from seeing the darn view count and upload date
on videos.

As usually, installed by running `./install tampermonkey` and then went
to: http://localhost/~wincent and navigating to the new script and
clicking it.