refactor: more tweaks for new version of git-cipher
In our "dotfiles" aspects we have some entries in "files" that are
really directories, and they contain encrypted files:

Directory:                     Contains:
----------                     ---------

.config/Raycast                .config/Raycast/Script Commands/emp.sh
                               .config/Raycast/Script Commands/the-hub.sh

.gitconfig.d                   .gitconfig.d/retiro

.zsh                           .zsh/exports.private
                               .zsh/hash.private

I am not too worried about the Raycast ones, because I'll always decrypt
on macOS, and that's the only place I use Raycast.

But ".gitconfig.d" could be a problem. If we include such a file without
decrypting it, Git will complain about bad config.

Likewise if we source the ".zsh" items and they are full of gibberish
because they're still encrypted, that will cause problems.

So, as mitigation:

1. In the ".zsh" files, we avoid sourcing them if we detect a magic
   header indicating that the file is still encrypted. I didn't want to
   call out to `git-cipher is-encrypted` from my shell startup files,
   because I want shell startup to be fast.

2. In the ".gitconfig.d" files, we'll skip over them if they aren't
   decrypted. Note that if you then do a `git-cipher lock`, the link
   will start pointing at ciphertext that Git will consider to be
   invalid garbage. In practice, we already have this problem with lots
   of other dotfiles; you're not really supposed to lock a repository
   that you're actively using to provide configuration to your machine.

In the "ssh" aspect,  we have similar stuff going on. The main
"~/.ssh/config" file is templated out, and we won't do that if it is not
decrypted. It itself includes files that are links to (hopefully)
decrypted files. We skip creating those links if things aren't
decrypted, with the same results: if you later lock, they will suddenly
become invalid, and you probably shouldn't do that.

Includes submodule update:

* vendor/git-cipher 35ab0b5...17e05c5 (5):
  > style: format more docs
  > feat(add): staged new files in Git index
  > docs: clean out old documentation
  > refactor: uncomment some code
  > refactor: remove dead code