chore: teach Prettier to ignore an encrypted JSON file
Because in CI it will be encrypted and will fail as seen here:

    https://github.com/wincent/wincent/runs/7933492379?check_suite_focus=true

like so:

    $ npx -y prettier --check "**/*.{js,json,md,ts}" "*.{md,ts}"
    Checking formatting...
    Error:  aspects/dotfiles/files/.github-handles.json: SyntaxError: Unexpected token (2:1)
    Error:    1 | magic = com.wincent.git-cipher
    Error:  > 2 | url = https://github.com/wincent/git-cipher/blob/v2.0.0-pre.2/PROTOCOL.md
    Error:      | ^
    Error:    3 | version = 1
    Error:    4 | algorithm = aes-256-cbc
    Error:    5 | filename = "aspects/dotfiles/files/.github-handles.json"
    All matched files use Prettier code style!
    error Command failed with exit code 2.

I could make a more complex solution here that changes the glob that
CI uses to select files for format checks (ie. based on the `CI`
environment variable), but this file is so trivial that I am choosing
the easy way out instead.