chore(nvim): prefer Ripgrep for Command-T
I need to figure out what I want to do in the long term because it is
probably something stupidly elaborate like:

1. Use Watchman if available, falling back to...
2. Ripgrep if Watchman is not available, falling back to...
3. Git if in a Git repo and Ripgrep is not available, failing back to...
4. The vanilla scanner.

Or something like that; maybe the order of 2 and 3 is swapped. Need to
think about it a bit more.

But in the meantime, Ripgrep is fast and if I want to find something
that Ripgrep usually wouldn't find, I can always run a different command
instead of hitting the mapping.