refactor(nvim): simplify separator handling
There were two inconsistencies here:

- One is that we were calling `dir_separator()` in one place, but not
  another.
- The other is that one place was carefully guarding against the
  possibility that `package.config` might not be set, but not in
  another.

Given that it evidently _is_ true (because it doesn't blow up in the
place where we're not doing the check), we can assume it's always true.
The Lua 5.1 docs don't mention it, but the 5.2 docs do:

- http://www.lua.org/manual/5.2/manual.html#pdf-package.config

and evidently LuaJIT must be following the same pattern.

Given that all that's left then is a call to `string.match()`, just
assign it to a module-local variable. There is no cost to pay at startup
for this, because this file is only ever loaded on demand when it is
actually needed.