refactor(nvim): remove unused parameter from `include_paths()`
`ext` is not used, so the call to `include_paths()` with "fnl" as the
second parameter isn't doing anything. Now, `package.path` is going to
look something like this (but all on one line, separated by ";"):

    ./?.lua
    /usr/local/share/luajit-2.1.0-beta3/?.lua
    /usr/local/share/lua/5.1/?.lua
    /usr/local/share/lua/5.1/?/init.lua
    /usr/local/share/lua/5.1/?.lua
    /usr/local/share/lua/5.1/?/init.lua

It is possible to configure Lua's `require()` to automatically compile
".fnl" files on demand:

- https://fennel-lang.org/api#use-luas-built-in-require-function

but I don't know whether there is any expectation that you should be
able to add "?.fnl" entries to your `package.path` or not. I presume you
probably can't, and that what's actually expected is that a ".lua" file
gets compiled at the same location. As such, we'd actually expect our
code here to check for ".fnl" files first, because the user probably
would rather see the source files than the compiled ones.

That's something to figure out if I ever become a Fennel user, or one
tells me about it. In the meantime, there's dead code to kill, so I'm
going to remove it rather than blindly "fixing" it to support a use case
I'm not even familiar with.