feat(nvim): add `safe-par` as wrapper around par
Because I find it annoying that I hit `gq` in Vim and my selection gets
unhelpfully replaced with:

    par error:
    Word too long: ...

par prints this out to standard out, so you can't hide it with a `2>
/dev/null` redirect. I haven't tested this, but my reading of the
man page suggests that it immediately aborts at this point, so even if
we could redirect the error message, you could still wind up with
something like this in your buffer:

   wrapped line
   wrapped line
   (omitted line due to error)
   (subsequent lines omitted due to early exit)

So, here we add the most inelegant of wrappers (inelegant because we
need not one but two temporary files) to shield us from this unwanted
behavior. In the first  we capture the input. In the second we capture
the output of `par`. If `par` succeeds, we return the output. Otherwise
we return the input.