refactor(codespaces): bring in .tmux-local.conf
This was in my private codespaces dotfiles repo, but that set-up
recently got broken by some upstream refactoring (the setup script
stopped running, which meant I had to rename my setup script, and that
in turn meant that symlinks stopped being created automatically).

So, as this file doesn't have anything sensitive in it, just bring it
over here instead (it doesn't need to be encrypted).