refactor(codespaces): don't die if sshd pidfile isn't present
More upstream changes, I suspect, as I am seeing errors like this on
initial bootstrap:

     pkill: pidfile not valid
     Try `pkill --help' for more information.

Logging into the VM later on, I can definitely see a pidfile there, and
the command works without issue:

    sudo pkill -HUP -F /var/run/sshd.pid

So, I'm going to assume that if the pidfile isn't present, the daemon is
restarting or something, in which case it will re-read the config file
anyway. Lame, but I don't see any other obvious choice at this point.