chore(nvim): turn on Sorbet by default
No need to do this on Codespaces because the ones I use already have it.
Can't do it on Darwin for annoying reasons noted in the comment (and
setting up a custom Ruby earlier in the `$PATH` seems fraught and maybe
not worth it). May still want to do it on non-Codespace Linux box, but
can cross that bridge later on.