chore: note what we _would_ do if we wanted to install Homebrew Ruby
So I had little play with this to see if it is worth it, and sadly, I
don't think it is.

The rationale: Sorbet requires at least Ruby 2.7.0, and macOS Ventura
(2022) ships with Ruby 2.6 (2018). Ruby 2.6 is old/unsupported, so
"overwriting" it seams reasonable.

The reality: my system has a fragile dependency on Ruby. Neovim depends
on Ruby, and I get Neovim from Homebrew. I need Ruby support in Neovim
for Command-T (well, not really, as I am using the Lua version now, but
I'd still like it to work). It is not clear to me whether installing
Neovim before/after installing Ruby will produce a working set-up, or
whether I can have them be somewhat independent.

So, I tried out the code in this commit, adding the new Ruby to the
start of the path. Neovim `:checkhealth` notes that the Command-T C
extension doesn't work. If I try to use it, I see this:

    command-t.vim could not load the C extension.
    Please see INSTALLATION and TROUBLE-SHOOTING in the help.
    Vim Ruby version: 2.6.10-p210
    Expected version: 2.6.8-p205
    For more information type:    :help command-t-ruby

which is actually super puzzling. I think this means that the extension
didn't load, and the version numbers are misleading about why. The
current OS (Ventura) is running 2.6.10, and that's why that shows up
even though it isn't first in the `$PATH` (when Neovim was built, maybe
it linked against a specific version, or maybe it didn't). When
Command-T was built, pre-Ventura, 2.6.8 was probably the current
version, so that's what's showing up in the message. In reality, Neovim
must be using not 2.6.10 but the 3.1.0 version coming from Homebrew,
which is why the extension isn't loading.

I could fix all this by rebuilding, but it doesn't feel particularly
robust. So, I put my `$PATH` back the way it was before, did a rebuild,
and everything works again.

The other problem. Even with the new version of Ruby, I can't run `srb`
in the main repo I work in because it dies looking for a million
uninstalled Gems. I don't want to (and perhaps can't) install those, and
much less on an ARM Mac. So, if `srb` won't run, Neovim can't use it to
provide LSP services.

Finally, the bootstrap process is annoyingly interdependent. For
example, I install Neovim and Ruby with Brew in the "homebrew" aspect.
Later, in the "ruby" aspect, I install gems, but for those to work, the
`$PATH` has to already be up-to-date. But the `$PATH` won't be
up-to-date because it was set up in the "dotfiles" aspect earlier on in
the process, but from an older shell, which means that the `$PATH` won't
become active until next time I open a terminal. I could fix this by
updating the path in the running install process, but that feels rather
bothersome.

Overall, I deem this too much effort and the juice is not worth the
squeeze.