feat(aur): add `yay` helper
Because `yay` will ask for sudo privileges on demand, and you
can't/shouldn't run it as root, we have to make sure the sudo timestamp
doesn't expire or isn't set by the time we run.

Will try this on a branch first as I am not sure it will work (not sure
if I have some restriction on where I can run `command` from).