feat(fig): handle "special" permissions
Handling may be describing it a little bit generously, as this is only
lightly tested, but...

On Codespace I am currently using I finally find myself in an
environment where "exotic" permissions are in use, which was causing my
`stat` calls to fall afoul of the `assert.mode` check which expected
modes to be things like "0755" but not "2755".

So, loosen the check, and things work. At the same time, I took the
opportunity to use a template literal type to replace my big long union:

- https://www.typescriptlang.org/docs/handbook/2/template-literal-types.html

Note that the docs say:

> We generally recommend that people use ahead-of-time generation for
> large string unions, but this is useful in smaller cases.

but... unless I find a big down side to getting this
auto-cross-multiplied `Mode` type via the template literal type, it sure
looks better.

Anyway, as I said, this is only lightly tested. I have no idea what will
happen if you start with a directory of type "2755" and I set it to
"0755" during a run. I expect it will remain "2755", which means that it
will try to set the permission on every single run. It is "idempotent"
only in the sense that the side effect it is trying to have isn't fully
taking place. But this does seem to work well enough to have unblocked
me on this Codespace at least.