fix(codespaces): undo `$PATH` bustage
I think `/etc/zsh/zlogin` has been on the main codespace image that I
work in for a while, but it is now overwriting `$PATH` unconditionally
(ie. not appending or prepending) in a way that blows away all my
carefully constructed `$PATH` shenanigans. Perhaps they started adding
`$PATH` to it only recently.

This is a problem because `zlogin` runs _after_ `zshrc`, as explained
here:

- https://unix.stackexchange.com/questions/71253/what-should-shouldnt-go-in-zshenv-zshrc-zlogin-zprofile-zlogout

Luckily, though, we can add a `~/.zlogin` and that will run even later
than `/etc/zsh/zlogin`, and give us a chance to patch things up. My
`append_to` and `prepend_to` helpers are clever enough to not add the
same path twice, so this should be relatively safe (ie. even though
we're setting up the `$PATH` in both `~/.zshrc` and `~/.zlogin`, we
should still wind up with a lean `$PATH`).