fix(nvim): fix "w" matching logic while setting `'formatprg'`
Intent here was to find a "w\d+" pattern in `'formatprg'` and change the
number to match current `'textwidth'`. If `'textwidth'` was 0, it would
be set to 9999 (the largest value that `par` supports).

ie. given a `'formatprg'` like:

    par rTbgqR B=.,\?_A_a_0 w100 Q=_s\>\|

and a `'textwidth'` of 72, it should update the `'formatprg'` to:

    par rTbgqR B=.,\?_A_a_0 w72 Q=_s\>\|

What was actually happening, however, is it would find the "w" in a
`'formatprg'` like this:

    /home/wincent/.zsh/bin/safe-par rTbgqR B=.,\?_A_a_0 Q=_s\>\|

and then fail to do anything. The "w" in "wincent" would match, but
there is no "w\d+" pattern, so no substitution would occur.

The first part of the fix then is to not match "w" in "wincent" but
rather look for:

    par .+w\d

ie. "par" followed by a space, then anything, then "w" and a digit. This
isn't totally immune to false positives (ie. you might have a
`'formatprg'` of "/home/par foow1/bin/par ...") but it should be robust
enough for practical purposes.

The second part of the fix is to adjust the `substitute` call
accordingly to not just replace "w\d+" but specifically to look for
"par" followed by space, followed by "w" followed by digits, and
surgically replace the "w" followed by digits bit only.

This probably used to work fine up until 79252479bff5074c7 ("feat(nvim):
add `safe-par` as wrapper around par", Sep 10, 2022), which changed my
`'formatprg'` from "par ..." to "/home/wincent/path/to/par ...".