chore(dotfiles): ignore some Raycast stuff
Seeing as a I just logged in to Raycast and used an invite link to get
access to the internal store, I know have an access token in here that I
don't want to commit to the repo (at least, I presume the `config.json`
file got added because of that).

Likewise, I have an extension now, downloaded and managed by Raycast
itself, so I don't want that committed either.