refactor(nvim): remove `on_visual=false` from `vim.highlight.on_yank()`
I've had it this way ever since 6b81b5aa9cd808d1 ("fix(vim): update
vim.highlight.on_yank to follow API changes", Jul 7, 2021). Before that,
I was achieving the same thing in a different way in ab263af7aabda917
("feat(vim): don't highlight yank if in VISUAL mode", Jun 3, 2020.).
IIRC, my rationale there was that the main reason I want `TextYankPost`
at all is to confirm for myself that I actually yanked what I meant to
(eg. when using a motion), and you don't need that in `VISUAL` mode
because you can, by definition, see what is selected.

However, right now, I have problems on this machine where yanked stuff
isn't making it into my system clipboard, and the lack of `TextYankPost`
is confusing my diagnosis (because I can't tell whether its because the
autocmd didn't run, or because it did run but something downstream of it
caused the text to not make it all the way to clipboard). So, just do it
all the time, for consistency.