chore(dotfiles): turn zsh syntax highlighting again
Back in f994c74eff617261b0726f3253c3b4fd320db918 ("zsh: try out syntax
highlighting", Jun 14, 2017) I wrote:

    Too strident for now, so I am committing this commented out. I will
    probably toggle it on and off for a bit to see whether I get used to
    it.

And I was thinking of deleting the submodule and removing it because
I've had it off for over 6 years now(!). But with my current color
scheme (base16 bright) it doesn't actually look that bad, so I'm going
to enable it. Ha. In addition it seems to fix a regression I've had with
zsh-autosuggestions starting in the last few days where possible
completions are not always rendered with dim highlighting.