perf(dotfiles): set up colors less aggressively
Instead of running the `tmux set` commands every single shell, only do
set-up if outside of tmux, and write a file that tmux will read at
startup.

This means the colors can get out of sync if you open multiple windows,
but I almost never do this and it makes the shell start up faster.

Next low-hanging fruit is a call to yarn that is showing up in the
profiling added in this commit.