fix: prepare to freshen base16 themes
Another episode in the hellscape that is Base16, this time, it's because
the maintainer of two of the themes in:

- https://github.com/chriskempson/base16-schemes-source

Has deleted their GitHub account. There are two PRs up to fix this by
removing the 404-ing links:

- https://github.com/chriskempson/base16-schemes-source/pull/7

or substituting them:

- https://github.com/chriskempson/base16-schemes-source/pull/5

Both of those have been sitting there since February, with no signs of
them being merged (and it is all a bit wild-west-y anyway with anybody
able to step up and claim that they maintain a popular scheme like
Solarized now).

There is a community "fork" of the schemes source repo, but that, too,
seems to be abandoned, having seen no commits since October 2018 (5
years ago):

- https://github.com/base16-community/base16-schemes-source

So, add a hack to allow me to use the last copy of the schemes that I
happen to have on disk on this machine. That will work until I change
machines, or until I grab a snapshot of them.

At some point I may just say "feck this", and pick a new, single scheme,
and stick with it.