chore(dotfiles): tweak delta settings
Make colors look better as per:

- https://github.com/tinted-theming/base16-shell

> Include delta.gitconfig in your Git config file i.e. ~/.gitconfig:
>
>     [delta]
>         syntax-theme = "ansi" # Use terminal colors

Also switch to side-by-side.

I rarely use delta, but when I do, it's to make sense of "ugly" diffs
like the one in the parent commit, and those are easier to read in
side-by-side mode.

Sadly, I can't figure out how to temporarily turn _off_ side-by-side. I
tried doing it with delta "features" (see `FEATURES` in `delta --help`)
but to no avail. As I don't use it much, I will live with it like this.